+++ messages/uk.json
@@ -24,6 +24,6 @@
   "notFoundPage": {},
   "buttons": {
     "moreAboutUs": "Більше про нас",
-    "callBack": "Передзвоніть мені"
+    "callBack": "Замовити зворотній дзвінок"
   }
 }
