+++ src/components/shared/header/CallBackButton.tsx
@@ -1,13 +1,12 @@
 import { PHONE_NUMBER } from "@/constants/constants";
 import { phoneRegex } from "@/regex/phoneRegex";
+import { useTranslations } from "next-intl";
 import Image from "next/image";
 
-export default function PhoneLink() {
+export default function CallBackButton() {
+  const t = useTranslations("buttons");
   return (
-    <a
-      href={`tel:+38${PHONE_NUMBER.replace(/\D/g, "")}`}
-      target="_blank"
-      rel="noopener noreferrer nofollow"
+    <button
       className="group flex items-center gap-x-1 outline-none text-18med laptop:hover:text-greenDark focus-visible:text-greenDark active:text-greenDark
        transition duration-300 ease-out"
     >
@@ -17,7 +16,10 @@ export default function PhoneLink() {
         width={40}
         height={40}
       />
-      {PHONE_NUMBER.replace(phoneRegex, "$1($2) $3 $4 $5")}
-    </a>
+      <div>
+        {PHONE_NUMBER.replace(phoneRegex, "$1($2) $3 $4 $5")}
+        <p className="text-12reg">{t("callBack").toLowerCase()}</p>
+      </div>
+    </button>
   );
 }
