+++ src/components/shared/header/headerMobTablet/BurgerMenuButton.tsx
@@ -19,14 +19,14 @@ export default function BurgerMenuButton({
       aria-label="open menu button"
       type="button"
       onClick={toggleHeaderMenuOpen}
-      className="group absolute top-2 right-4 z-50 w-14 h-14 px-[10px] py-[15px] outline-none before:content-['']
+      className="group absolute top-2 right-4 z-[60] w-14 h-14 px-[10px] py-[15px] outline-none before:content-['']
          before:absolute before:-z-10 before:top-0 before:left-0 before:rounded-full before:size-14 before:opacity-0 before:transition 
          before:duration-1000 before:ease-out active:before:opacity-100 focus-visible:before:opacity-100 laptop:hover:before:opacity-100 
          before:bg-greenGradient before:blur-[4px]"
     >
       <div className="w-full h-full rotate-0 ">
         <span
-          className={`z-50 block absolute w-full h-[1.8px] rounded-md bg-greenDark rotate-0 
+          className={`block absolute w-full h-[1.8px] rounded-md bg-greenDark rotate-0 
                 transition duration-[600ms] ease-out ${
                   isHeaderMenuOpened
                     ? firstLineTransformStyles
@@ -34,7 +34,7 @@ export default function BurgerMenuButton({
                 }`}
         ></span>
         <span
-          className={`z-50 block absolute left-0 w-full h-[1.8px] rounded-md bg-greenDark rotate-0 
+          className={`block absolute left-0 w-full h-[1.8px] rounded-md bg-greenDark rotate-0 
                 transition duration-[600ms] ease-out ${
                   isHeaderMenuOpened
                     ? secondLineTransformStyles
@@ -42,7 +42,7 @@ export default function BurgerMenuButton({
                 }`}
         ></span>
         <span
-          className={`z-50 block absolute left-0 w-full h-[1.6px] rounded-md bg-greenDark  rotate-0  
+          className={`block absolute left-0 w-full h-[1.6px] rounded-md bg-greenDark  rotate-0  
                 transition duration-[600ms] ease-out ${
                   isHeaderMenuOpened
                     ? thirdLineTransformStyles
