+++ src/components/shared/header/headerMobTablet/HeaderMobTablet.tsx
@@ -6,9 +6,9 @@ import { useLocale } from "next-intl";
 import { Locale } from "@/types/locale";
 import BurgerMenuButton from "./BurgerMenuButton";
 import HeaderMobTabBgImages from "./HeaderMobTabBgImages";
-import PhoneLink from "../PhoneLink";
-import BurgerMenu from "./BurgerMenu";
+import BurgerMenu from "./burgerMenu/BurgerMenu";
 import Backdrop from "../../backdrop/Backdrop";
+import CallBackButton from "../CallBackButton";
 
 interface HeaderMobTabletProps {
   color?: "beige" | "white";
@@ -34,7 +34,7 @@ export default function HeaderMobTablet({
         </div>
         <LogoLink />
         <div className="hidden tab:flex">
-          <PhoneLink />
+          <CallBackButton />
         </div>
         <div className="size-14">
           <BurgerMenuButton
